profile_diffs.py: script to compare Go coverage profiles, outputs new blocks
This python script accepts two paths to Go coverage profiles and outputs the
new code blocks that were hit in the second profile and not the first. If the
script is used on both permutations of coverage profiles, it's possible to see
whether coverage increased or decreased fairly easily. The script should be
adapted in the future to:
 - use branch coverage (see something like: https://github.com/junhwi/gobco/)
 - automatically parse this output data and tell how many new branches were hit
   and how many old branches were not hit.
 - compare the number of hit counters instead of "hit" vs. "not hit".

Help from ChatGPT 4o